refactor: move resoucesNames.ts outside state directory
