+++ src/resoucesNames.ts
